Tratando as exceptions
Adicionei o tratamento básico das exceptions nos seguintes casos:
-Email duplicado (post, put);
-Id não encontrado (get, put, delete);

coloquei para retornar uma mensagem personalizada na exception "NoSuchElementException".

+++ back/src/main/java/com/studentscool/StudentsCool/adapters/in/web/EscolaController.java
@@ -2,13 +2,17 @@ package com.studentscool.StudentsCool.adapters.in.web;
 
 import com.studentscool.StudentsCool.application.domain.Escola;
 import com.studentscool.StudentsCool.application.ports.in.EscolasUseCases;
+import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.NoSuchElementException;
 
 @RestController
 @RequestMapping("/escolas")
@@ -19,9 +23,16 @@ public class EscolaController {
     private EscolasUseCases escolasUseCases;
 
     @PostMapping()
-    public Escola criarEscola(@RequestBody Escola escola) {
-        Escola resultado = escolasUseCases.PostEscola(escola);
-        return resultado;
+    public ResponseEntity<?> criarEscola(@RequestBody Escola escola) {
+        try {
+            Escola resultado = escolasUseCases.PostEscola(escola);
+            return ResponseEntity.status(HttpStatus.CREATED).body(escola);
+        } catch (DataIntegrityViolationException e) {
+            Map<String, String> errorResponse = new HashMap<>();
+            errorResponse.put("Mensagem", "Erro: Email duplicado.");
+            return ResponseEntity.status(HttpStatus.CONFLICT).body(errorResponse);
+        }
+
     }
 
     @GetMapping
@@ -30,23 +41,64 @@ public class EscolaController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Escola> getEscolaById(@PathVariable Long id) {
-        Escola escola = escolasUseCases.getEscolaById(id);
-        return ResponseEntity.ok().body(escola);
+    public ResponseEntity<?> getEscolaById(@PathVariable(value = "id") Long id) {
+        try {
+            Escola escola = escolasUseCases.getEscolaById(id);
+            return ResponseEntity.status(HttpStatus.OK).body(escola);
+        } catch (NoSuchElementException e) {
+            Map<String, String> errorResponse = new HashMap<>();
+            errorResponse.put("Mensagem", e.getMessage());
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
+        } catch (Exception e) {
+            Map<String, String> errorResponse = new HashMap<>();
+            errorResponse.put("Mensagem", "Erro interno no servidor.");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
+        }
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Escola> atualizarEscola(@PathVariable(value = "id") Long id,
+    public ResponseEntity<?> atualizarEscola(@PathVariable(value = "id") Long id,
                                                   @RequestBody Escola escolaDetails) {
-        Escola updateEscola = escolasUseCases.updateEscola(escolaDetails, id);
-        return ResponseEntity.ok(updateEscola);
+        try {
+            Escola updateEscola = escolasUseCases.updateEscola(escolaDetails, id);
+            return ResponseEntity.ok(updateEscola);
+        } catch (NoSuchElementException e) {
+            Map<String, String> errorResponse = new HashMap<>();
+            errorResponse.put("Mensagem", e.getMessage());
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
+        } catch (DataIntegrityViolationException e) {
+            Map<String, String> errorResponse = new HashMap<>();
+            errorResponse.put("Mensagem", "Erro: Email duplicado.");
+            return ResponseEntity.status(HttpStatus.CONFLICT).body(errorResponse);
+        } catch (Exception e) {
+            Map<String, String> errorResponse = new HashMap<>();
+            errorResponse.put("Mensagem", "Erro interno no servidor.");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
+        }
     }
 
     @DeleteMapping("/{id}")
-    public Map<String, Boolean> deleteEscola(@PathVariable(value = "id") Long id) {
-        escolasUseCases.deleteEscola(id);
-        Map<String, Boolean> response = new HashMap<>();
-        response.put("Deletado", Boolean.TRUE);
-        return response;
+    public ResponseEntity<Map<String, String>> deleteEscola(@PathVariable(value = "id") Long id) {
+        try {
+            List<Escola> todasEscolas = escolasUseCases.getAllEscolas();
+
+            boolean idExiste = todasEscolas.stream().anyMatch(escola -> escola.getId().equals(id));
+
+            if (!idExiste) {
+                Map<String, String> errorRepsponse = new HashMap<>();
+                errorRepsponse.put("Mensagem", "Erro: Escola com o ID " + id + " não foi encontrada");
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorRepsponse);
+            }
+
+            escolasUseCases.deleteEscola(id);
+
+            Map<String, String> response = new HashMap<>();
+            response.put("message", "Escola com o ID: " + id + " foi deletada com sucesso");
+            return ResponseEntity.ok(response);
+        } catch (Exception e) {
+            Map<String, String> errorResponse = new HashMap<>();
+            errorResponse.put("Mensagem", "Erro interno no servidor: " + e.getMessage());
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
+        }
     }
 }
